requirements.txt
websockets==12.0
pandas==2.2.2
numpy==1.26.4
aiosqlite==0.19.0
python-binance==1.0.19
pytz==2024.1
Flask==3.0.3